Fixed the 'non-design documents' issue
It's natural that the connector throw a 'non-design documents' exception when the Cloudant database is empty or just contains '_design' docs. The issue is that the connector might throw this exception as well when the database contains both '_design' docs and non-design docs.
To fix this issue, changed the logic of 'getMany' method to:
1. Get the docs by the Cloudant '_all_docs' API.
2. Try to get the docs again if 1 returns a empty list, but this time it will use 'startkey' to filter the '_design' documents.

BugzID: 59236